Update README.md
The website offers a user-friendly experience with a pop-up login feature on the home page for easy access. Users can navigate through a carousel slick and use the property search option in the menu to find specific listings. Additionally, there is a cart functionality available in the menu for convenient transactions. To learn more about the platform, users can visit the 'About Us' page, and for any inquiries, they can access the 'Contact Us' page.